Move default definitions to defaults module
