+++ crates/config/src/defaults.rs
@@ -18,6 +18,13 @@ use std::path::PathBuf;
 
 use crate::unresolved_config::LogLevel;
 
+const DEFAULT_HASHES_PER_ROUND: usize = 1024;
+const DEFAULT_ASYNC_TO_SYNC_QUEUE_SIZE: usize = 1;
+const DEFAULT_SYNC_TO_ASYNC_QUEUE_SIZE: usize = 1;
+
+const DEFAULT_UTILITY_QUEUE_SIZE: usize = 100;
+const DEFAULT_FACADE_QUEUE_SIZE: usize = 100;
+
 pub(crate) fn default_log_level() -> LogLevel {
     LogLevel::Error
 }
@@ -32,4 +39,24 @@ pub(crate) fn default_state_path() -> PathBuf {
 
 pub(crate) fn default_msr_enabled() -> bool {
     false
+}
+
+pub(crate) fn default_hashes_per_round() -> usize {
+    DEFAULT_HASHES_PER_ROUND
+}
+
+pub(crate) fn default_async_to_sync_queue_size() -> usize {
+    DEFAULT_ASYNC_TO_SYNC_QUEUE_SIZE
+}
+
+pub(crate) fn default_sync_to_async_queue_size() -> usize {
+    DEFAULT_SYNC_TO_ASYNC_QUEUE_SIZE
+}
+
+pub(crate) fn default_utility_queue_size() -> usize {
+    DEFAULT_UTILITY_QUEUE_SIZE
+}
+
+pub(crate) fn default_facade_queue_size() -> usize {
+    DEFAULT_FACADE_QUEUE_SIZE
 }
